requirements.txt
fastapi[all]
uvicorn
pymongo
redis
python-dotenv==0.19.2